+++ Project1/GameObject.h
@@ -2,7 +2,7 @@
 class GameObject
 {
 public:
-	float location[3];
+	double location[3];
 	float rotation[3] = { 0,0,0 };
 	GameObject(int, int, int);
 	~GameObject();
